// MegabinFE/src/app/services/model/getScheduleContract.ts
/**
 * MySmartCloset API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DayOfWeek } from './dayOfWeek';
import { Frequency } from './frequency';


export interface GetScheduleContract { 
    id: string;
    frequency: Frequency;
    dayOfWeek: DayOfWeek;
    startingDate: string;
    lastCollected?: string | null;
    active: boolean;
    approvedExternally: boolean;
    addressesId: string;
}
export namespace GetScheduleContract {
}