// MegabinFE/src/app/services/model/updateScheduleContract.ts
/**
 * MySmartCloset API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DayOfWeek } from './dayOfWeek';
import { Frequency } from './frequency';

export interface UpdateScheduleContract {
  contractId?: string;
  frequency?: Frequency;
  dayOfWeek?: DayOfWeek;
  active?: boolean;
  approvedExternally?: boolean;
}
export namespace UpdateScheduleContract {}